data-collection/requirements.txt
aiohttp==3.8.4
aioselenium==0.0.1
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
asynctest==0.13.0
attrs==23.1.0
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.6.0
multidict==6.0.4
typing-extensions==4.5.0
yarl==1.9.2
zipp==3.15.0
